now looking at your screen should feel better

+++ srcs/signals/handler_signal.c
@@ -6,7 +6,7 @@
 /*   By: sylducam <sylducam@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/05 09:57:19 by tglory            #+#    #+#             */
-/*   Updated: 2021/12/23 16:07:08 by sylducam         ###   ########.fr       */
+/*   Updated: 2021/12/23 16:24:34 by sylducam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,10 +22,9 @@ void ft_sigquit(void *master)
 
 	if (!save)
 		save = master;
-	else if (save->pid > 0)
+	else if (save->pid > 0 && g_ctrl_c == 0) // keep 2nd condition ? I don't think so
 	{
-		dprintf(1, "\nsyltest pid > 0\n"); // to delete
-		g_ctrl_c = 1;
+		g_ctrl_c = 1; // keep it ?
 		kill(save->pid, SIGQUIT);
 		printf("\n");
 		// rl_on_new_line();
@@ -45,9 +44,8 @@ void ft_sigint(void *master)
 	{
 		if (save->pid == -1)
 		{
-			dprintf(1, "\nsyltest pid == -1\n"); // to delete
 			printf("\n");
-			g_ctrl_c = 1;
+			g_ctrl_c = 1; // keep it ?
 			// rl_on_new_line();
 			rl_replace_line("", 0);
 			// rl_redisplay();
@@ -56,10 +54,8 @@ void ft_sigint(void *master)
 		}
 		if (save->pid)
 		{
-			dprintf(1, "\nsyltest pid > -1\n"); // to delete
 			kill(save->pid, SIGINT);
-			g_ctrl_c = 1;
-			dprintf(1, "\nX syltest g_ctrl_c = %d\n", g_ctrl_c); // to delete
+			g_ctrl_c = 1; // keep it ?
 			printf("\n");
 			rl_on_new_line();
 			rl_replace_line("", 0);
